Move `video.h` to `src/misc`
