+++ src/audio/mixer.cpp
@@ -29,11 +29,11 @@
 #include "hardware/memory.h"
 #include "hardware/pic.h"
 #include "hardware/timer.h"
-#include "hardware/video/video.h"
 #include "midi/midi.h"
 #include "misc/cross.h"
 #include "misc/notifications.h"
 #include "misc/tracy.h"
+#include "misc/video.h"
 #include "utils/checks.h"
 #include "utils/math_utils.h"
 #include "utils/rwqueue.h"
