+++ src/cpu/cpu.cpp
@@ -17,9 +17,9 @@
 #include "gui/mapper.h"
 #include "gui/titlebar.h"
 #include "hardware/pic.h"
-#include "hardware/video/video.h"
 #include "lazyflags.h"
 #include "misc/support.h"
+#include "misc/video.h"
 #include "shell/command_line.h"
 #include "utils/math_utils.h"
 #include "utils/string_utils.h"
