+++ src/debugger/debugger.cpp
@@ -30,10 +30,10 @@
 #include "hardware/pic.h"
 #include "hardware/port.h"
 #include "hardware/timer.h"
-#include "hardware/video/video.h"
 #include "misc/cross.h" //snprintf
 #include "misc/std_filesystem.h"
 #include "misc/support.h"
+#include "misc/video.h"
 #include "shell/shell.h"
 #include "utils/string_utils.h"
