+++ src/dos/programs/boot.cpp
@@ -16,10 +16,10 @@
 #include "gui/titlebar.h"
 #include "hardware/dma.h"
 #include "hardware/input/mouse.h"
-#include "hardware/video/video.h"
 #include "hardware/virtualbox.h"
 #include "hardware/vmware.h"
 #include "ints/bios_disk.h"
+#include "misc/video.h"
 #include "more_output.h"
 #include "utils/fs_utils.h"
 #include "utils/string_utils.h"
