+++ src/dosbox.cpp
@@ -35,13 +35,13 @@
 #include "hardware/port.h"
 #include "hardware/timer.h"
 #include "hardware/video/reelmagic/reelmagic.h"
-#include "hardware/video/video.h"
 #include "hardware/video/voodoo.h"
 #include "ints/int10.h"
 #include "midi/midi.h"
 #include "misc/cross.h"
 #include "misc/support.h"
 #include "misc/tracy.h"
+#include "misc/video.h"
 #include "shell/shell.h"
 #include "utils/math_utils.h"
