+++ src/gui/private/sdlmain.h
@@ -19,7 +19,7 @@
 
 #include "gui/common.h"
 #include "gui/render.h"
-#include "hardware/video/video.h"
+#include "misc/video.h"
 #include "utils/fraction.h"
 #include "utils/rect.h"
