+++ src/gui/render.cpp
@@ -19,8 +19,8 @@
 #include "gui/mapper.h"
 #include "gui/render.h"
 #include "hardware/video/vga.h"
-#include "hardware/video/video.h"
 #include "misc/support.h"
+#include "misc/video.h"
 #include "shell/shell.h"
 #include "utils/fraction.h"
 #include "utils/math_utils.h"
