+++ src/gui/sdlmain.cpp
@@ -55,11 +55,11 @@
 #include "hardware/pic.h"
 #include "hardware/timer.h"
 #include "hardware/video/vga.h"
-#include "hardware/video/video.h"
 #include "ints/int10.h"
 #include "misc/cross.h"
 #include "misc/notifications.h"
 #include "misc/tracy.h"
+#include "misc/video.h"
 #include "shell/command_line.h"
 #include "utils/checks.h"
 #include "utils/fs_utils.h"
