+++ src/gui/shader_manager.cpp
@@ -15,8 +15,8 @@
 #include <SDL.h>
 
 #include "dosbox.h"
-#include "hardware/video/video.h"
 #include "misc/cross.h"
+#include "misc/video.h"
 #include "utils/checks.h"
 #include "utils/math_utils.h"
 #include "utils/string_utils.h"
