+++ src/hardware/input/mouse_config.cpp
@@ -9,8 +9,8 @@
 #include "config/config.h"
 #include "config/setup.h"
 #include "gui/common.h"
-#include "hardware/video/video.h"
 #include "misc/support.h"
+#include "misc/video.h"
 #include "utils/checks.h"
 #include "utils/math_utils.h"
 #include "utils/string_utils.h"
