+++ src/hardware/video/reelmagic/video_mixer.cpp
@@ -17,7 +17,7 @@
 
 #include "config/setup.h"
 #include "gui/render_scalers.h" //SCALER_MAXWIDTH SCALER_MAXHEIGHT
-#include "hardware/video/video.h"
+#include "misc/video.h"
 #include "utils/checks.h"
 #include "utils/rgb565.h"
