+++ src/hardware/video/vga.cpp
@@ -10,9 +10,9 @@
 
 #include "gui/common.h"
 #include "hardware/pic.h"
-#include "hardware/video/video.h"
 #include "ints/int10.h"
 #include "misc/logging.h"
+#include "misc/video.h"
 #include "utils/math_utils.h"
 #include "utils/string_utils.h"
 
@@ -87,80 +87,6 @@ void VGA_DetermineMode(void) {
 	case 7: VGA_SetMode(M_LIN24); break;
 	case 13:VGA_SetMode(M_LIN32);break;
 	}
-}
-
-const char* to_string(const GraphicsStandard g)
-{
-	switch (g) {
-	case GraphicsStandard::Hercules: return "Hercules";
-	case GraphicsStandard::Cga: return "CGA";
-	case GraphicsStandard::Pcjr: return "PCjr";
-	case GraphicsStandard::Tga: return "Tandy";
-	case GraphicsStandard::Ega: return "EGA";
-	case GraphicsStandard::Vga: return "VGA";
-	case GraphicsStandard::Svga: return "SVGA";
-	case GraphicsStandard::Vesa: return "VESA";
-	default: assertm(false, "Invalid GraphicsStandard"); return "";
-	}
-}
-
-const char* to_string(const ColorDepth c)
-{
-	switch (c) {
-	case ColorDepth::Monochrome: return "monochrome";
-	case ColorDepth::Composite: return "composite";
-	case ColorDepth::IndexedColor2: return "2-colour";
-	case ColorDepth::IndexedColor4: return "4-colour";
-	case ColorDepth::IndexedColor16: return "16-colour";
-	case ColorDepth::IndexedColor256: return "256-colour";
-	case ColorDepth::HighColor15Bit: return "15-bit high colour";
-	case ColorDepth::HighColor16Bit: return "16-bit high colour";
-	case ColorDepth::TrueColor24Bit: return "24-bit true colour";
-	default: assertm(false, "Invalid ColorDepth"); return "";
-	}
-}
-
-const char* to_string(const PixelFormat pf)
-{
-	switch (pf) {
-	case PixelFormat::Indexed8: return "Indexed8";
-	case PixelFormat::RGB555_Packed16: return "RGB555_Packed16";
-	case PixelFormat::RGB565_Packed16: return "RGB565_Packed16";
-	case PixelFormat::BGR24_ByteArray: return "BGR24_ByteArray";
-	case PixelFormat::BGRX32_ByteArray: return "BGRX32_ByteArray";
-	default: assertm(false, "Invalid pixel format"); return {};
-	}
-}
-
-uint8_t get_bits_per_pixel(const PixelFormat pf)
-{
-	return enum_val(pf);
-}
-
-// Return a human-readable description of the video mode, e.g.:
-//   - "CGA 640x200 16-colour text mode 03h"
-//   - "EGA 640x350 16-colour graphics mode 10h"
-//   - "VGA 720x400 16-colour text mode 03h"
-//   - "VGA 320x200 256-colour graphics mode 13h"
-//   - "VGA 360x240 256-colour graphics mode"
-//   - "VESA 800x600 256-colour graphics mode 103h"
-std::string to_string(const VideoMode& video_mode)
-{
-	const char* mode_type = (video_mode.is_graphics_mode ? "graphics mode"
-	                                                     : "text mode");
-
-	const auto mode_number = (video_mode.is_custom_mode
-	                                  ? ""
-	                                  : format_str(" %02Xh",
-	                                                  video_mode.bios_mode_number));
-
-	return format_str("%s %dx%d %s %s%s",
-	                     to_string(video_mode.graphics_standard),
-	                     video_mode.width,
-	                     video_mode.height,
-	                     to_string(video_mode.color_depth),
-	                     mode_type,
-	                     mode_number.c_str());
 }
 
 const char* to_string(const VGAModes mode)
