+++ src/hardware/video/vga_draw.cpp
@@ -17,8 +17,8 @@
 #include "gui/render_scalers.h"
 #include "hardware/pic.h"
 #include "hardware/video/reelmagic/reelmagic.h"
-#include "hardware/video/video.h"
 #include "ints/int10.h"
+#include "misc/video.h"
 #include "utils/bitops.h"
 #include "utils/math_utils.h"
 #include "utils/mem_unaligned.h"
