+++ src/misc/video.h
